+++ src/constants/paths.ts
@@ -1,7 +1,9 @@
 export enum Path {
+  LEGACY = '/legacy',
   LEVERAGE = '/leverage',
   PRESALES = '/presales',
   PRODUCTS = '/products',
   PRT_STAKING = '/prt-staking',
   TRADE = '/swap',
+  YIELD = '/yield',
 }
