+++ src/feature-flags.ts
@@ -1,3 +1,6 @@
 // Example: enable everywhere except on prod deployment
 // export const isFeatureEnabled = () =>
 //   process.env.NEXT_PUBLIC_VERCEL_ENV !== 'index-app-prod'
+
+export const isYieldPageEnabled = () =>
+  process.env.NEXT_PUBLIC_VERCEL_ENV !== 'index-app-prod'
